tweaks to work with neo4j-admin import
Still has a race condition that for some reason might drop the first 4
bytes of a file. Bizarre! Workaround is using some garbage 4 bytes in a
header file. /shrug

This does mean compressed files don't work. I wonder if this is because
of some magic byte reading taking place in Java? No idea!